Split ExecuteDynamicCode god classes under 500 lines
Extract Class on DynamicCodeLiteralHoister, SourceShaper, and SharedRoslynCompilerWorkerHost so each file has one clear responsibility while keeping Rewrite/Analyze/WrapIfNeeded and worker host APIs behaviorally unchanged.
