+++ Packages/src/Editor/FirstPartyTools/ExecuteDynamicCode/DynamicCompilation/DynamicCodeLiteralHoister.cs
@@ -11,20 +11,6 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
     internal static class DynamicCodeLiteralHoister
     {
         internal const string LiteralParameterPrefix = "__uloop_literal_";
-        private static readonly Dictionary<char, char> RegularStringEscapes = new()
-        {
-            { '\'', '\'' },
-            { '"', '"' },
-            { '\\', '\\' },
-            { '0', '\0' },
-            { 'a', '\a' },
-            { 'b', '\b' },
-            { 'f', '\f' },
-            { 'n', '\n' },
-            { 'r', '\r' },
-            { 't', '\t' },
-            { 'v', '\v' }
-        };
 
         public static HoistedLiteralRewriteResult Rewrite(string source)
         {
@@ -35,27 +21,27 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
 
             while (index < source.Length)
             {
-                if (TryCopyInterpolatedStringLiteral(source, rewrittenSource, ref index))
+                if (DynamicCodeLiteralSyntaxScanner.TryCopyInterpolatedStringLiteral(source, rewrittenSource, ref index))
                 {
                     continue;
                 }
 
-                if (TryCopyVerbatimStringLiteral(source, rewrittenSource, ref index))
+                if (DynamicCodeLiteralSyntaxScanner.TryCopyVerbatimStringLiteral(source, rewrittenSource, ref index))
                 {
                     continue;
                 }
 
-                if (TryCopyCharLiteral(source, rewrittenSource, ref index))
+                if (DynamicCodeLiteralSyntaxScanner.TryCopyCharLiteral(source, rewrittenSource, ref index))
                 {
                     continue;
                 }
 
-                if (TryCopyLineComment(source, rewrittenSource, ref index))
+                if (DynamicCodeLiteralSyntaxScanner.TryCopyLineComment(source, rewrittenSource, ref index))
                 {
                     continue;
                 }
 
-                if (TryCopyBlockComment(source, rewrittenSource, ref index))
+                if (DynamicCodeLiteralSyntaxScanner.TryCopyBlockComment(source, rewrittenSource, ref index))
                 {
                     continue;
                 }
@@ -91,7 +77,7 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
 
                 if (scopeTracker.ShouldSuppressLiteralHoisting)
                 {
-                    if (TryCopyRegularStringLiteral(source, rewrittenSource, ref index))
+                    if (DynamicCodeLiteralSyntaxScanner.TryCopyRegularStringLiteral(source, rewrittenSource, ref index))
                     {
                         continue;
                     }
@@ -124,512 +110,6 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 declarationLines);
         }
 
-        private static bool TryCopyInterpolatedStringLiteral(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            int start = index;
-            if (!TryAdvanceInterpolatedStringLiteral(source, ref index))
-            {
-                return false;
-            }
-
-            rewrittenSource.Append(source, start, index - start);
-            return true;
-        }
-
-        private static bool TryCopyLineComment(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            if (index + 1 >= source.Length || source[index] != '/' || source[index + 1] != '/')
-            {
-                return false;
-            }
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                rewrittenSource.Append(current);
-                index++;
-                if (current == '\n')
-                {
-                    break;
-                }
-            }
-
-            return true;
-        }
-
-        private static bool TryCopyBlockComment(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            if (index + 1 >= source.Length || source[index] != '/' || source[index + 1] != '*')
-            {
-                return false;
-            }
-
-            rewrittenSource.Append(source[index]);
-            rewrittenSource.Append(source[index + 1]);
-            index += 2;
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                rewrittenSource.Append(current);
-                index++;
-                if (current == '*' && index < source.Length && source[index] == '/')
-                {
-                    rewrittenSource.Append('/');
-                    index++;
-                    break;
-                }
-            }
-
-            return true;
-        }
-
-        private static bool TryCopyVerbatimStringLiteral(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            if (source[index] != '@'
-                || index + 1 >= source.Length
-                || source[index + 1] != '"')
-            {
-                return false;
-            }
-
-            int start = index;
-            index += 2;
-
-            while (index < source.Length)
-            {
-                if (source[index] != '"')
-                {
-                    index++;
-                    continue;
-                }
-
-                if (index + 1 < source.Length && source[index + 1] == '"')
-                {
-                    index += 2;
-                    continue;
-                }
-
-                index++;
-                rewrittenSource.Append(source, start, index - start);
-                return true;
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static bool TryAdvanceInterpolatedStringLiteral(string source, ref int index)
-        {
-            int start = index;
-            if (!TryMatchInterpolatedStringStart(source, ref index, out bool isVerbatim))
-            {
-                return false;
-            }
-
-            int interpolationDepth = 0;
-            while (index < source.Length)
-            {
-                if (interpolationDepth > 0)
-                {
-                    interpolationDepth = AdvanceInterpolatedExpressionSegment(
-                        source,
-                        ref index,
-                        interpolationDepth);
-                    continue;
-                }
-
-                InterpolatedStringAdvanceResult advanceResult =
-                    AdvanceInterpolatedStringContentSegment(source, index, isVerbatim);
-                index = advanceResult.Index;
-                interpolationDepth = advanceResult.InterpolationDepth;
-                if (advanceResult.Completed)
-                {
-                    return true;
-                }
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static int AdvanceInterpolatedExpressionSegment(
-            string source,
-            ref int index,
-            int interpolationDepth)
-        {
-            if (TryAdvanceInterpolatedExpressionToken(source, ref index))
-            {
-                return interpolationDepth;
-            }
-
-            char expressionCharacter = source[index];
-            if (expressionCharacter == '{')
-            {
-                index++;
-                return interpolationDepth + 1;
-            }
-
-            if (expressionCharacter == '}')
-            {
-                index++;
-                return interpolationDepth - 1;
-            }
-
-            index++;
-            return interpolationDepth;
-        }
-
-        private static InterpolatedStringAdvanceResult AdvanceInterpolatedStringContentSegment(
-            string source,
-            int index,
-            bool isVerbatim)
-        {
-            char current = source[index];
-            if (current == '{')
-            {
-                return index + 1 < source.Length && source[index + 1] == '{'
-                    ? new InterpolatedStringAdvanceResult(index + 2, 0, false)
-                    : new InterpolatedStringAdvanceResult(index + 1, 1, false);
-            }
-
-            if (current == '}')
-            {
-                return index + 1 < source.Length && source[index + 1] == '}'
-                    ? new InterpolatedStringAdvanceResult(index + 2, 0, false)
-                    : new InterpolatedStringAdvanceResult(index + 1, 0, false);
-            }
-
-            if (!isVerbatim && current == '\\')
-            {
-                int escapedIndex = index;
-                AdvanceEscapedLiteralSequence(source, ref escapedIndex);
-                return new InterpolatedStringAdvanceResult(escapedIndex, 0, false);
-            }
-
-            if (current != '"')
-            {
-                return new InterpolatedStringAdvanceResult(index + 1, 0, false);
-            }
-
-            if (isVerbatim && index + 1 < source.Length && source[index + 1] == '"')
-            {
-                return new InterpolatedStringAdvanceResult(index + 2, 0, false);
-            }
-
-            return new InterpolatedStringAdvanceResult(index + 1, 0, true);
-        }
-
-        private static bool TryMatchInterpolatedStringStart(
-            string source,
-            ref int index,
-            out bool isVerbatim)
-        {
-            isVerbatim = false;
-
-            if (source[index] == '$')
-            {
-                if (index + 1 < source.Length && source[index + 1] == '"')
-                {
-                    index += 2;
-                    return true;
-                }
-
-                if (index + 2 < source.Length && source[index + 1] == '@' && source[index + 2] == '"')
-                {
-                    isVerbatim = true;
-                    index += 3;
-                    return true;
-                }
-
-                return false;
-            }
-
-            if (source[index] != '@')
-            {
-                return false;
-            }
-
-            if (index + 2 < source.Length && source[index + 1] == '$' && source[index + 2] == '"')
-            {
-                isVerbatim = true;
-                index += 3;
-                return true;
-            }
-
-            return false;
-        }
-
-        private static bool TryAdvanceInterpolatedExpressionToken(string source, ref int index)
-        {
-            if (TryAdvanceInterpolatedStringLiteral(source, ref index))
-            {
-                return true;
-            }
-
-            if (TryAdvanceVerbatimStringLiteral(source, ref index))
-            {
-                return true;
-            }
-
-            if (TryAdvanceRegularStringLiteral(source, ref index))
-            {
-                return true;
-            }
-
-            if (TryAdvanceCharLiteral(source, ref index))
-            {
-                return true;
-            }
-
-            if (TryAdvanceLineComment(source, ref index))
-            {
-                return true;
-            }
-
-            if (TryAdvanceBlockComment(source, ref index))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        private static bool TryAdvanceLineComment(string source, ref int index)
-        {
-            if (index + 1 >= source.Length || source[index] != '/' || source[index + 1] != '/')
-            {
-                return false;
-            }
-
-            index += 2;
-            while (index < source.Length && source[index] != '\n')
-            {
-                index++;
-            }
-
-            if (index < source.Length)
-            {
-                index++;
-            }
-
-            return true;
-        }
-
-        private static bool TryAdvanceBlockComment(string source, ref int index)
-        {
-            if (index + 1 >= source.Length || source[index] != '/' || source[index + 1] != '*')
-            {
-                return false;
-            }
-
-            index += 2;
-            while (index < source.Length)
-            {
-                if (source[index] == '*' && index + 1 < source.Length && source[index + 1] == '/')
-                {
-                    index += 2;
-                    return true;
-                }
-
-                index++;
-            }
-
-            return true;
-        }
-
-        private static bool TryAdvanceVerbatimStringLiteral(string source, ref int index)
-        {
-            if (source[index] != '@'
-                || index + 1 >= source.Length
-                || source[index + 1] != '"')
-            {
-                return false;
-            }
-
-            int start = index;
-            index += 2;
-
-            while (index < source.Length)
-            {
-                if (source[index] != '"')
-                {
-                    index++;
-                    continue;
-                }
-
-                if (index + 1 < source.Length && source[index + 1] == '"')
-                {
-                    index += 2;
-                    continue;
-                }
-
-                index++;
-                return true;
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static bool TryAdvanceRegularStringLiteral(string source, ref int index)
-        {
-            if (source[index] != '"')
-            {
-                return false;
-            }
-
-            if (index > 0 && (source[index - 1] == '@' || source[index - 1] == '$'))
-            {
-                return false;
-            }
-
-            int start = index;
-            index++;
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                if (current == '\\')
-                {
-                    AdvanceEscapedLiteralSequence(source, ref index);
-                    continue;
-                }
-
-                if (current == '"')
-                {
-                    index++;
-                    return true;
-                }
-
-                index++;
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static bool TryAdvanceCharLiteral(string source, ref int index)
-        {
-            if (source[index] != '\'')
-            {
-                return false;
-            }
-
-            int start = index;
-            index++;
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                if (current == '\\')
-                {
-                    AdvanceEscapedLiteralSequence(source, ref index);
-                    continue;
-                }
-
-                if (current == '\'')
-                {
-                    index++;
-                    return true;
-                }
-
-                index++;
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static bool TryCopyCharLiteral(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            if (source[index] != '\'')
-            {
-                return false;
-            }
-
-            int start = index;
-            index++;
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                if (current == '\\')
-                {
-                    AdvanceEscapedLiteralSequence(source, ref index);
-                    continue;
-                }
-
-                if (current == '\'')
-                {
-                    index++;
-                    rewrittenSource.Append(source, start, index - start);
-                    return true;
-                }
-
-                index++;
-            }
-
-            index = start;
-            return false;
-        }
-
-        private static bool TryCopyRegularStringLiteral(
-            string source,
-            StringBuilder rewrittenSource,
-            ref int index)
-        {
-            if (source[index] != '"')
-            {
-                return false;
-            }
-
-            if (index > 0 && (source[index - 1] == '@' || source[index - 1] == '$'))
-            {
-                return false;
-            }
-
-            int start = index;
-            index++;
-
-            while (index < source.Length)
-            {
-                char current = source[index];
-                if (current == '\\')
-                {
-                    AdvanceEscapedLiteralSequence(source, ref index);
-                    continue;
-                }
-
-                if (current == '"')
-                {
-                    index++;
-                    rewrittenSource.Append(source, start, index - start);
-                    return true;
-                }
-
-                index++;
-            }
-
-            index = start;
-            return false;
-        }
-
         private static bool TryHoistRegularStringLiteral(
             string source,
             StringBuilder rewrittenSource,
@@ -654,7 +134,7 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 char current = source[index];
                 if (current == '\\')
                 {
-                    AdvanceEscapedLiteralSequence(source, ref index);
+                    DynamicCodeRegularStringLiteralUnescaper.AdvanceEscapedLiteralSequence(source, ref index);
                     continue;
                 }
 
@@ -662,7 +142,7 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
                 {
                     index++;
                     string literalToken = source.Substring(start, index - start);
-                    if (!TryUnescapeRegularStringLiteral(literalToken, out string value))
+                    if (!DynamicCodeRegularStringLiteralUnescaper.TryUnescapeRegularStringLiteral(literalToken, out string value))
                     {
                         rewrittenSource.Append(literalToken);
                         return true;
@@ -765,218 +245,6 @@ namespace io.github.hatayama.UnityCliLoop.FirstPartyTools
         private static string CreateParameterName(int index)
         {
             return $"{LiteralParameterPrefix}{index}";
-        }
-
-        private static void AdvanceEscapedLiteralSequence(string source, ref int index)
-        {
-            index++;
-            if (index >= source.Length)
-            {
-                return;
-            }
-
-            char escape = source[index];
-            index++;
-
-            switch (escape)
-            {
-                case 'u':
-                    index = Math.Min(source.Length, index + 4);
-                    break;
-                case 'U':
-                    index = Math.Min(source.Length, index + 8);
-                    break;
-                case 'x':
-                    index = AdvanceVariableLengthHexDigits(source, index, 4);
-                    break;
-            }
-        }
-
-        private static int AdvanceVariableLengthHexDigits(string value, int index, int maxDigits)
-        {
-            int digitsConsumed = 0;
-            while (index < value.Length && digitsConsumed < maxDigits && IsHexDigit(value[index]))
-            {
-                index++;
-                digitsConsumed++;
-            }
-
-            return index;
-        }
-
-        private static bool TryUnescapeRegularStringLiteral(string token, out string value)
-        {
-            string inner = token.Substring(1, token.Length - 2);
-            StringBuilder unescaped = new(inner.Length);
-
-            for (int index = 0; index < inner.Length; index++)
-            {
-                char current = inner[index];
-                if (current != '\\')
-                {
-                    unescaped.Append(current);
-                    continue;
-                }
-
-                if (!TryAppendRegularStringEscape(inner, ref index, unescaped))
-                {
-                    value = null;
-                    return false;
-                }
-            }
-
-            value = unescaped.ToString();
-            return true;
-        }
-
-        private static bool TryAppendRegularStringEscape(
-            string inner,
-            ref int index,
-            StringBuilder unescaped)
-        {
-            if (index + 1 >= inner.Length)
-            {
-                return false;
-            }
-
-            index++;
-            char escape = inner[index];
-            if (RegularStringEscapes.ContainsKey(escape))
-            {
-                unescaped.Append(RegularStringEscapes[escape]);
-                return true;
-            }
-
-            return TryAppendComplexRegularStringEscape(inner, ref index, unescaped, escape);
-        }
-
-        private static bool TryAppendComplexRegularStringEscape(
-            string inner,
-            ref int index,
-            StringBuilder unescaped,
-            char escape)
-        {
-            switch (escape)
-            {
-                case 'u':
-                    return TryAppendUnicodeEscape(inner, ref index, unescaped);
-                case 'U':
-                    return TryAppendUtf32Escape(inner, ref index, unescaped);
-                case 'x':
-                    return TryAppendVariableLengthHexEscape(inner, ref index, unescaped);
-                default:
-                    return false;
-            }
-        }
-
-        private static bool TryAppendUnicodeEscape(
-            string inner,
-            ref int index,
-            StringBuilder unescaped)
-        {
-            if (!TryParseHexDigits(inner, ref index, 4, out int unicodeValue))
-            {
-                return false;
-            }
-
-            unescaped.Append((char)unicodeValue);
-            return true;
-        }
-
-        private static bool TryAppendUtf32Escape(
-            string inner,
-            ref int index,
-            StringBuilder unescaped)
-        {
-            if (!TryParseHexDigits(inner, ref index, 8, out int codePoint) ||
-                !IsValidUtf32CodePoint(codePoint))
-            {
-                return false;
-            }
-
-            unescaped.Append(char.ConvertFromUtf32(codePoint));
-            return true;
-        }
-
-        private static bool TryAppendVariableLengthHexEscape(
-            string inner,
-            ref int index,
-            StringBuilder unescaped)
-        {
-            if (!TryParseVariableLengthHexDigits(inner, ref index, out int variableLengthValue))
-            {
-                return false;
-            }
-
-            unescaped.Append((char)variableLengthValue);
-            return true;
-        }
-
-        private static bool TryParseHexDigits(string value, ref int index, int digitCount, out int parsedValue)
-        {
-            int start = index + 1;
-            if (start + digitCount > value.Length)
-            {
-                parsedValue = 0;
-                return false;
-            }
-
-            string hex = value.Substring(start, digitCount);
-            if (!int.TryParse(hex, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out parsedValue))
-            {
-                return false;
-            }
-
-            index += digitCount;
-            return true;
-        }
-
-        private static bool TryParseVariableLengthHexDigits(string value, ref int index, out int parsedValue)
-        {
-            int start = index + 1;
-            int end = AdvanceVariableLengthHexDigits(value, start, 4);
-            if (end == start)
-            {
-                parsedValue = 0;
-                return false;
-            }
-
-            string hex = value.Substring(start, end - start);
-            if (!int.TryParse(hex, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out parsedValue))
-            {
-                return false;
-            }
-
-            index = end - 1;
-            return true;
-        }
-
-        private static bool IsValidUtf32CodePoint(int value)
-        {
-            return value >= 0
-                && value <= 0x10FFFF
-                && (value < 0xD800 || value > 0xDFFF);
-        }
-
-        private static bool IsHexDigit(char value)
-        {
-            return (value >= '0' && value <= '9')
-                || (value >= 'a' && value <= 'f')
-                || (value >= 'A' && value <= 'F');
-        }
-
-        private readonly struct InterpolatedStringAdvanceResult
-        {
-            public InterpolatedStringAdvanceResult(int index, int interpolationDepth, bool completed)
-            {
-                Index = index;
-                InterpolationDepth = interpolationDepth;
-                Completed = completed;
-            }
-
-            public int Index { get; }
-            public int InterpolationDepth { get; }
-            public bool Completed { get; }
         }
     }
